Move LINKS to Admin/constants
